+++ app/src/main/kotlin/ir/vmessenger/navigation/HomeGraph.kt
@@ -39,7 +39,6 @@ private fun rememberHomeNavigation(navController: NavHostController): HomeNaviga
             onNavigateToIdentity = { navController.navigate(VmRoute.Identity) },
             onNavigateToNodes = { navController.navigate(VmRoute.Nodes) },
             onNavigateToAbout = { navController.navigate(VmRoute.About) },
-            onNavigateToUpdate = { navController.navigate(VmRoute.Update) },
             onNavigateToBlockedContacts = { navController.navigate(VmRoute.BlockedContacts) },
             onNavigateToActivityLog = { navController.navigate(VmRoute.ActivityLog) },
             onNavigateToDebug = { navController.navigate(VmRoute.Debug) },